Enable readiness check in the config_reload of the core_dump and config check (#14142)
Description of PR
Summary:
core_dump_and_config_check will issue config_reload to the duthosts after test module finished and if there's coredump or config_db changes.
For now, it simply issues the config_reload command then returns and finishes.
For the testbeds that require more time to be ready, it possibly fails on the sanity check of the next test module.
Hence enable the readiness check of the config_reload.

Adjust the order of teardown of sanity_check and core_dump fixture

co-authorized by: jianquanye@microsoft.com

Change-Id: I2cbeff915be4f00f9eda60ef88a9203005c346df