Fix issue RM#4139550
This issue is fixed by community PR https://github.com/sonic-net/sonic-mgmt/pull/13400. But it's too hard to have a clean cherry-pick of the PR because our local code has not been synchronized with community for too long time. In the next merge, there should be conflict, we can just drop this CR by that time.

Change-Id: I600ed43f88fd5b623d8528df6678aa6a3284d214