Add support for skip tests on the smartswitch DPU (#15699)
- What is the motivation for this PR?
Add support for skipping tests on the smartswitch DPU

- How did you do it?
There are 2 main changes in this PR:

When getting the basic facts of the DUT in the conditional mark plugin, use the dut name in the --host-pattern parameter instead of the first dut defined in the testbed file. By this change, we will be able to get the dpu facts when running the test only on the dpu.
Update the cache logic to save the cache with the dut name. So that in a same regression run, we are able to save and get cache for both the switch and the dpu.

- How did you verify/test it?
Run regression in our internal regression. It works fine.

- Any platform specific information?
This change is only for the smartswitch testbed, and it will not harm the regular testbeds.

Change-Id: I7c34399a3d17e4942a781ceb41f801d291e03b9d