[nvos] make action_reboot func wait for system is ready
the reload func that is eventually called by operations that cause reboot, is doing the following after the reboot operation:
1. wait for ping down
2. wait for ping up
3. if topology_obj specified, wait for system is ready in serial
4. try ssh
5. so on ...

before the change - if not explicitly passing topology_obj to this func, it's not waiting for system is ready, and may fail when engine tries to ssh before system is ready.

with this change - if not explicitly passing topology_obj to this func, it's taking it from TestToolkit, and make it wait to system is ready too.

Change-Id: I321b3ac95cfd9ffa5eda50629f0dcd00dfe24a20