Fix PTF TACACS always use default password issue (#15802)
What is the motivation for this PR?
The code for setting up TACACS on PTF host has an issue of always using default password.

How did you do it?
The original intention is to get an alternate password. If the alternate password cannot be found, then use default password.

However, the step of using default password always override the alternate password.

This change added code to store the alternate password in variable tacacs_user_passwd. Then the override logic won't happen if alternate password is found.

Change-Id: Ia935b6974b3ec26276d980cd9f9553636f08b0b9