Add check to see if bgp sessions are up before proceeding with the tests (#15921)
Description of PR
After acl reboot, we check for critical process status and interface status but do not check to see if all BGP sessions are up or not. Sometimes testcases fail because packet is not received, this is because BGP sessions are not yet up and fail the ACL tests for the wrong reasons.

This check will make sure that BGP sessions are up before continuing with the ACL tests

Instead of checking only for T1 and M0 profile, checking it for all profiles

Summary:
Fixes # (issue)

Approach
What is the motivation for this PR?
After acl reboot, we check for critical process status and interface status but do not check to see if all BGP sessions are up or not. Sometimes testcases fail because packet is not received, this is because BGP sessions are not yet up and fail the ACL tests for the wrong reasons.

How did you do it?
This check will make sure that BGP sessions are up before continuing with the ACL tests

How did you verify/test it?
On T2 profile for Cisco 8808 chassis

Any platform specific information?
No

co-authorized by: jianquanye@microsoft.com

Change-Id: I73f68e9da5502b8ca53488ecad7ed98ee9ffb94e