Disable pfcwd for pfc_lossless. (#15887)
Description of PR
The pfc lossless test assume pfcwd is disabled. But this test: test_pfc_pause_multi_lossless_prio is not provided the fixture: "disable_pfcwd" which is used by all other tests in the same script. This causes this test to fail since the pfcwd is triggered and packets are dropped.

Approach
What is the motivation for this PR?
this test is failing with this error:
Failed: Total TX bytes   841563217920 should be smaller than DUT buffer size 67108864

How did you do it?
Disable pfcwd during the test using the fixture, that is already used by other tests in the same script.

How did you verify/test it?
Ran it on my TB:

co-authorized by: jianquanye@microsoft.com

Change-Id: I414e4213fb6c9e8c1dcf5d28d1050525cb973c4d