Add remaining control plane test scripts to onboarding PR checkers (#15907)
What is the motivation for this PR?
Some control plane test scripts had not yet been integrated into the PR checkers. In this PR, we onboard these scripts to ensure they are included in the PR testing process.

How did you do it?
In this PR, we onboard these scripts to ensure they are included in the PR testing process.

How did you verify/test it?

Change-Id: Id5aecd1463cfbf7b16c9b2d9c8de2a1fdc0a1271