chore: skip if no console port info on dut (#15979)
Description of PR
Summary:
Fixes # (issue) 304717824

Type of change
For devices that doesnt have console port (i.e Cisco 8800 console only exists in RP), we should skip this test for this DUT. Otherwise we are going to have KeyError: 'ManagementIp'

Approach
What is the motivation for this PR?
How did you do it?
We skip if the key does not exists. If key does not exists then the graph will be empty. For example:

 "dut-lc1-1": {},
How did you verify/test it?
verified on physical testbed

Change-Id: Iee7568388e0b282e1022bcb59f72c21fe15e3b0b
Signed-off-by: Austin Pham <austinpham@microsoft.com>