Increase the timeout for fwutil test
This is a temp fix which will not be up streamed. Currently it could take 30 mins to burn the FPGA on a bobcat system, but HW team says they will introduce a turbo mode in the FPGA burning to speed up. Need to revert this CR when the turbo mode is available.

Change-Id: Idd0207e99c1196635673ee7a7284e1899cd0314e