Add marker validation in pre-test stage of PR testing.  (#15877)
What is the motivation for this PR?
In our proposed PR testing model, we use pytest.mark.topology to identify and collect relevant test scripts. To ensure all scripts are properly marked, particularly newly added ones, we introduced a checker in the pre-test stage to verify the presence and correctness of these markers.

How did you do it?
To ensure all scripts are properly marked, particularly newly added ones, we introduced a checker in the pre-test stage to verify the presence and correctness of these markers.

How did you verify/test it?
Tested by pipeline itself. If a script lack the marker, it will fail the whole pipeline.

Change-Id: Ifad243d4a6c91efee9fd7c36bcdc660562d3dc24