[Chassis][voq] TC test_voq_chassis_app_db_consistency.py Modification Due to Lag ID Set Changes (#16116)
What is the motivation for this PR?
Changes by the Functionality sonic-net/sonic-buildimage#20369 of Lag ID assignment.
The TC Expectation and assertion is changed

How did you do it?
Ignoring the SYSTEM_LAG_ID_SET to be same as in case pre-dump.
But rather be assigned from SYSTEM_LAG_IDS_FREE_LIST in order.
Added a sanity of lag_id_set to ensure the functionality of PR sonic-net/sonic-buildimage#20369.

How did you verify/test it?
Tested and Verified on a T2 VOQ Chassis

Change-Id: Ia1082e171c2bf7b9689b6f677a9dcd956674b5f4