[Mellanox] Fix test_shared_headroom_pool_configure issue (#13135)
For mellanox device, the headroom size is related to the speed and cable length. In community setup, some cable length is 300m in config, so when the speed is over 400G, the headroom size will exceed the limitation. So for
test_shared_headroom_pool_configure, we change the cable length to a smaller when speed is equal or greater 400G and the cable length is over 200m.

Change-Id: I4386beae3a1844b71a8a719fbad3471011abda7a