Fix fabric_cli_and_db so it only tests on connected links. (#16309)
Description of PR
Summary: Fix the fabric_cli_and_db test so it only tests connected links.
Fixes # (issue)

The fabric_cli_and_db failed in cases where the test selected a link that was not connected. This issue occurred because the portList set was not reset for each chip, allowing links that were connected on other chips to be tested, even though they were not connected on the chip being tested. In this PR, the portList is reset for each chip, and the 'show fabric reachability' command is also run on each chip. As a result, the test now only choose connected links on the chip under test.

Type of change
 Bug fix

Change-Id: I9dd94f1efe447a60676b878391d1196f95e56252