Support qos test in KVM testbed and skip fanout config and traffic test (#16157)
What is the motivation for this PR?
Elastictest performs well in distribute running PR test in multiple KVMs, which support us to add more test scripts to PR checker.
But some traffic test can't be tested on KVM platform, we need to skip traffic test if needed

How did you do it?
This PR adds qos tests to the KVM-based PR test framework with the following scope and modifications:

Excludes fanout switch-related configurations, which are not applicable in the KVM test environment.
Traffic tests have been intentionally skipped due to the limitations of running traffic in the KVM environment.
How did you verify/test it?

Change-Id: Iff58f624c9ac6907ffdd48ae497c8241aed2a733