sonic-mgmt: Fix namespace issues for qos tests on T2 single ASIC (#15708)
We are seeing UnboundLocalError when running sonic-mgmt tests against
a single-ASIC linecard:

```
UnboundLocalError: local variable 'dst_sys_port_id' referenced before assignment
```

Upon further investigation, this was determined to be happening
because a previous attempt to fix this issue (PR #13700) completely
omitted the ASIC prefix, but the entries in SYSTEM_PORT in config_db
do have an Asic0 prefix even on a single ASIC DUT.

Resolve this by specifically adding the Asic0 prefix in the case of a
single-ASIC T2 DUT, instead of leaving the prefix out.

Tested by manually running qos tests on a T2 single ASIC DUT with
these changes.

Change-Id: Ibb4109a066f7cc73302f9a032f26737a3c034bfb