Fix test_console_escape in new version of netmiko
1. Root cause
In newer versions of netmiko, the "send_command" method will first try to match the command itself, so it will try to match "\x03"
When we send "\x03" sent to the terminal, it displays as "^C" in the terminal, so in the output we see the string form: '^C\r\n\r\n...'
- "\x03" is a single-byte control character (ASCII value 3)
- "^C" is a combination of two characters (^ and C)
When we read the output from the terminal, we get the display form "^C", not the original control character "\x03". They are not matched, so case fail

2. Solution
Avoid using "send_command" to send the interrupt character. Instead, use "write_channel" to send the interrupt signal directly, and then use "read_until_pattern" to match the expected output

Current version:
- paramiko: 2.9.5
- netmiko: 3.2.0

Change-Id: I24835b65d8bbf7038519e0c9da743d98a8fcd5b6