Calculate diff accuratelly for PRs using `git merge base` (#16377)
What is the motivation for this PR?
Previously, we used the command git diff origin/master HEAD to get the differences between the current PR and the target master branch. However, this approach is unreliable because any changes made to the remote branch can alter the diff, which is not the intended behavior. In this PR, we address this issue by using git merge-base origin/master HEAD to determine the common ancestor between the current PR and the target branch. We then calculate the diff using git diff <ancestor> HEAD to accurately capture the changes in the current PR relative to the target branch.

How did you do it?
In this PR, we address the issue by using git merge-base origin/master HEAD to determine the common ancestor between the current PR and the target branch. We then calculate the diff using git diff <ancestor> HEAD to accurately capture the changes in the current PR relative to the target branch.

Change-Id: If8f6943278e3fc5d0a132c71a0edcb586753ef5b