chore: disable parallel run for ro disk test (#16510)
Description of PR
Disable parallel run for tacacs/test_ro_disk.py test.

Summary:
Fixes # (issue) Microsoft ADO 30848681

Approach
What is the motivation for this PR?
The tacacs/test_ro_disk.py test is using a special reboot process on the DUT and if we run this test with parallel run enabled, the LCs might not be ready after Supervisor's reboot. Therefore, we need to disable parallel run for this test.

co-authorized by: jianquanye@microsoft.com

Change-Id: If19f94584b43de45d5e7cbba6d22701d8e8c0f16