Add test_gnoi_system.py with a simple test for System.Time. (#16334)
Description of PR
Create a new file for testing GNOI System API and bootstrap it with a simple test for System.Time

Approach
What is the motivation for this PR?
Create a new file for testing GNOI System API and bootstrap it with a simple test for System.Time

How did you do it?
Create a new file for testing GNOI System API and bootstrap it with a simple test for System.Time

How did you verify/test it?
Run on KVM

Change-Id: Id158dec03f3d101b661aa1c94bd2c6c2f576999e