Fix skip condition issue
When branch name is not a official name, the branch match will not be exact as the current logic. E.g. when the fixed version is master_SN5640_RC.28-5efa37b9e_Internal, but the tested version is SONiC.master.628-8e246d26b_Internal,
they are not viewed as a same branch. So add dot to the tested branch string, and then check if the new string exist in the fixed version, it can resolve the issue.

Change-Id: I59f806a2d8ed797172942c800e1d791d74fa860b