Fix multiple failures in KillProcess test on KVM. (#16303)
Description of PR
Summary:

Gracefully skip the test for killprocess if the process does not exists, instead of failing it. Some process such as telemetry is not always there on KVM testbed and the validation of killprocess should not depend on it.
Avoid killing swss and snmp process as it leads to errors on KVM testbed:isAutoNegEnabled: Failed to get port AutoNeg status for port pid:1000000000021
Fixes #16238

Approach
The test tries to kill a process then restart it but the API throws an exception if the process does not exist.
Here we gracefully catch the exception and skip the test.

What is the motivation for this PR?
Fix the test for KVM.

Change-Id: Ie9d8a8c18454e0a31991d1fe52f9646d058521fb