Update percentage threshold based on DUT memory (#16671)
Summary
For Nokia-7215 which has only 3GB memory, diff percentage threshold 8% is not enough.
In this PR, I updated the threshold as below:
* DUT memory in [4GB, +inf): 4%  <-- Same as current impl.
* DUT memory in (3GB, 4GB] : 8%  <-- Arista-7050QX still fall in this range (as current impl)
* DUT memory in (2GB, 3GB] : 10% <-- Nokia-7215 will fall in this range
* DUT memory in (0, 2GB]   : 12% <-- Same as current impl.
Enhance the test failure message to print more information. Below are the new failure message

What is the motivation for this PR?
For Nokia-7215 which has only 3GB memory, diff percentage threshold 8% is not enough.
Enhance the test failure message to print more information.

How did you verify/test it?
Verified on Nokia 7215 M0 testbed.

Change-Id: I39dd30f33497b9df3fa0865957e7c1174fca0d68