Don't save config after copying over the backup files. (#16691)
Description of PR
Summary:
In the function: save_restore_config() in test_multidut_pfcwd_basic_with_snappi.py, the teardown section copies back the backup files to /etc/sonic/. But the next line is to save config, which is not right. This PR fixes this issue.

Approach
What is the motivation for this PR?
The backup files contain the original configs for the DUT, but the config save command will overwrite them. In teardown, we want the original configs back, not the running configs.

How did you do it?
Remove the config-save command in teardown.

How did you verify/test it?
Ran it on my TB. I checked if the running config is different from saved config, and verified the new code works correctly.

@sdszhang , @auspham : For viz. This might potentially fix the issues in pfcwd_basic/reboot cases.

co-authorized by: jianquanye@microsoft.com

Change-Id: Id17b889eeee211f2c4110aff68f331628edbc67d