Skip kubesonic test case for public branch (#16684)
What is the motivation for this PR?
The k8s feature in not enabled for public branches. So the test case should be skipped for public branches

How did you do it?
Check the image type and skip the test case if it is public branch

How did you verify/test it?
Yes

Any platform specific information?
No

Supported testbed topology if it's a new test case?
Not new case

Change-Id: I01eaf104fb6d23764be979e85d417c931776eef3